Many Bug Fixes
Several of the else statements were misaligned and had to be fixed. A "while" loop was added since there were issues with python not checking the code state over and over (it would just print the same welcome message over and over). Some text was moved over from Story.py to avoid a looped import. I also forgot to assign input() to a variable so that I could check if the user inputted what was required. Some state changes were fixed as well (during the last fight).